refactor: Move CameraPosition into model package
Moves the CameraPosition model object into the map model and updates it
to use Bounds instead of the GMS LatLngBounds.
