+++ ground/src/main/java/com/google/android/ground/ui/home/mapcontainer/LocationOfInterestRepositionView.kt
@@ -20,7 +20,7 @@ import com.google.android.ground.R
 import com.google.android.ground.databinding.MapMoveLocationOfInterestLayoutBinding
 import com.google.android.ground.rx.RxAutoDispose
 import com.google.android.ground.ui.common.AbstractView
-import com.google.android.ground.ui.map.CameraPosition
+import com.google.android.ground.model.map.CameraPosition
 import com.google.android.ground.ui.map.MapFragment
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.WithFragmentBindings
