+++ ground/src/main/java/com/google/android/ground/ui/home/mapcontainer/MapContainerViewModel.kt
@@ -28,6 +28,7 @@ import com.google.android.ground.model.geometry.Coordinate
 import com.google.android.ground.model.geometry.Point
 import com.google.android.ground.model.locationofinterest.LocationOfInterest
 import com.google.android.ground.model.locationofinterest.LocationOfInterestType
+import com.google.android.ground.model.map.CameraPosition
 import com.google.android.ground.model.map.MapLocationOfInterest
 import com.google.android.ground.repository.LocationOfInterestRepository
 import com.google.android.ground.repository.OfflineAreaRepository
